+++ src/Security/AuthProvider.php
@@ -12,9 +12,11 @@ class AuthProvider
     private bool $isAnonymous;
     private ?UuidInterface $userUuid;
     private ?UuidInterface $tokenUuid = null;
+    private ?string $hashedToken = null;
 
     public function __construct(
         private ParameterBagInterface $bag,
+        private TokenGenerator $tokenGenerator,
         private Server500LogicExceptionFactory $server500LogicExceptionFactory
     ) {
         $anonymousUserUuid = $this->bag->get('anonymousUserUUID');
@@ -25,13 +27,28 @@ class AuthProvider
         $this->isAnonymous = true;
     }
 
+    public function getRedisTokenKeyFromHashedToken(string $hashedToken): string
+    {
+        return sprintf(
+            'token:%s',
+            $hashedToken
+        );
+    }
+
+    public function getRedisTokenKeyFromRawToken(string $rawToken): string
+    {
+        return $this->getRedisTokenKeyFromHashedToken($this->tokenGenerator->hashToken($rawToken));
+    }
+
     public function setUserAndToken(
         UuidInterface $userUuid = null,
         UuidInterface $tokenUuid = null,
+        string $hashedToken = null,
         bool $isAnonymous = false
     ): self {
         $this->userUuid = $userUuid;
         $this->tokenUuid = $tokenUuid;
+        $this->hashedToken = $hashedToken;
         $this->isAnonymous = $isAnonymous;
 
         return $this;
@@ -45,6 +62,11 @@ class AuthProvider
     public function getUserUuid(): ?UuidInterface
     {
         return $this->userUuid;
+    }
+
+    public function getHashedToken(): ?string
+    {
+        return $this->hashedToken;
     }
 
     public function getTokenUuid(): ?UuidInterface