+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add feature test for GET `/token` endpoint
+- Add feature test for POST `/token` endpoint
+- Add feature test for DELETE `/token` endpoint
 ### Changed
+- Increase reference dataset version to 0.0.8.
 - **Switch license to GPL-3.0-only.**
 
 ## 0.0.27 - 2023-09-02